refactor brush and add logic
